pyproject.toml
[project]
name = "nats"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "nats-py>=2.11.0",
]
